Classe BLEService : ajout d'une classe de callback pour les opérations liées à la localisation et modification de l'implémentation du service.
Classe Localisation : ajout de la connection avec le BLEService.

+++ main/java/com/stuffinder/webservice/NetworkService.java
@@ -985,12 +985,12 @@ public class NetworkService implements NetworkServiceInterface {
     }
 
     @Override
-    public int getLastTagsUpdateTime() throws NetworkServiceException {
+    public long getLastTagsUpdateTime() throws NetworkServiceException {
         return 0;
     }
 
     @Override
-    public int getLastProfilesUpdateTime() throws NetworkServiceException {
+    public long getLastProfilesUpdateTime() throws NetworkServiceException {
         return 0;
     }
